feat(caritas-theme): moved legal pages from extensions to core
