+++ src/components/legalPageWrapper/legalPageWrapper.styles.scss
